Change project name and entry file, add src folder
Our projects name looked confusing so I changed the capitalization of a letter to help with its readability.
Added a src folder as webpack will look for a specific file to bundle from/for and created a src folder as the modules will exist in that directory and later result in building a dist folder with end user files placed into it.