Add webpack-merge, scripts, config files
Webpack-merge allows us to use slightly different config files for webpack depending on if we're going to run webpack with our code for development purposes or production purposes/testing.
Scripts allow us to call on these separate configurations of our webpack.
And as already explained somewhat via what webpack-merge does we now have multiple webpack versions and in these files the differences are stated and we merge in the common settings as well.