Change getProjects, add starting projectsArray,etc
First off we start this iife module pattern by setting an array as value to projectsArray if it did not hold an array prior, which will let us just add projects thereafter without worrying about the array.
Secondly I changed the getProjects function to pull the currently saved value of projectsArray out and convert it to js code from json string, then depending on how many projects are in that array we either produce a semi neat output or we tell the user there are no projects saved at this time.
I also changed the addProject and deleteProject to correctly parse the JSON data pulled from projectsArray however I'm trying to use an identifier to delete projects but that property is not being set up correctly so will have to look at the project.js further to fix that.