Add createNoteItem method, change method args
As I previously mentioned I was going to change the methods that deleted array elements to use index values this way they'll rely on a unique value which is dynamically set when actions change these arrays, and this will correlate with dataset attributes on ui elements and we can use these functions with ease there as well.
I have also created a createNoteItem method as that is very useful and not sure why I didn't before, another learning moment in regards to planning pre coding.