Remove identifier prop, change method and arg name
Once again the identifier value acquisition was not working as intended and have reverted to using index positions.
I've also changed the name of the method that allows us to change a property of a Project to a different value and changed its argument name to versions less verbose and more to the point.