Change deleteProject to use index, rename params
Like the other modules leading into this module, we are switching our methods that delete specific elements to use an index position instead of any specific property value as this will help us to delete a unique element in an array and not the first one with x or y name property value.
I've also renamed some parameters to help with describing what the method does and works with.