Create deleteProjects method, change names, etc
Inside the getProjects method I had a condition that hinged on when projects array had a length greater than one, however if there are no elements in that array then there is no length, therefore I changed the code to hinge on length or no length instead, and it still logs the projects from localStorage or it tells the user that there aren't any projects saved so far.
I also changed the name of addProject to createProject, it just sounds better to create a project, although we are adding a new project to an array, semantics.
Added deleteProjects which is just a renamed localStorage.clear call but just wanted to keep calls to the program namespace and not have to rely on using localStorage direct invocations.
And of course added the new method to be returned to the object made by this iife module pattern which gets imported into index.js.