Change deleteProjects to set empty item
Before this change we used the clear method on localStorage, but we weren't setting an array as a value after the clear, so any further actions would meet errors.
This change means that when a user deletes their projects array rather than wiping it, it gets overwritten instead with a new empty array.