Add projects expression, change methods
I'm shifting this application to use the browser more than before, and what I mean by that is that I was trying to just deal with localStorage via JSON in many of the methods, but I now see that's not going to work.
First I've added a variable which exists from the start of the application that'll hold all of the project elements converted to javascript.
I've renamed  getProjects logProjects as getProjects felt like acquiring them which is now automatically done anyways.
I've changed the methods to use the projects variable to make changes to projects and then push those changes to the local storage.