Add reviver and replacer arguments to JSON methods
The replacer and reviver methods arguments are important to us as they are what allows us to save a function as a string and to change a string into a function.
Changes in the files other than todoApp.js includes moving the class methods into the constructor function to create own methods in the instances of those classes, thus allowing us to save the methods within each object. This is important because when we stringify an object made from a class or constructor it loses hidden properties because when it gets converted back from JSON into a pojo then its just an object with no connections to anything, unless of course we stringify those as well and thus make them explicit, but no thanks.