Add Project methods, add Todo prop, add pseudocode
Added a complete property to Todo instances, which allows us to keep track of completed todos and gives us the ability to not show todos when focusing on what's needed to complete at a specific time, but later allow us to show a user all the work they've completed and somewhat goals succeeded.
Filled in logic for Project methods.
I'm going to need to call the save projects function within each of these methods that makes a change to a project or todo or checklist, but that'll be done later, I've made a pseudocode comment as a reminder. I also added a comment to the load function pseudocode.