Add functionality for todoApp
I recreated the todoApp module, which is an iife factory function which by invoking allows us to export its functions and variables to one location our index.js.
I was able to not use a replacer or reviver in my JSON functions when saving and loading data to and from localStorage.
My biggest snag was in the loadStorage function which not only parses JSON data from localStorage and converts it into not only javascript code but into new instances of classes, but when I went to overwrite the projects variable I was mistakenly passing returnedProjects in reference, and not actually making copies, and this was fixed by using a for of loop and pushing the objects in.