Fix assignment of projects in startANew
At this time we call splice on projects variable and how we use it it wipes out all of the contents of that array, but what I had there earlier was resetting the deleted items into projects. This is because when you call splice splice manipulates the array object itself, you don't need to reassign, if you do reassign it saves the values you just removed back into the array.