Declare variables, declare modal to create project
Declared some of the starting variables so I can target elements that I need to add content to.
Created a modal that hides by default and unhides , by changing its size, when create project button has been clicked and hides when saved.
Now that I think of it I need a cancel button as well, so will build that next.