Add a cancel button to the create project modal
This button allows users to leave the modal and wipe anything they've input, obviously we want to give them the ability to not make a new project and close the modal if they need/want to.
I also got rid of some id values I had dynamically added to elements, I thought that I needed to give elements an id to work with them after we append the modal into the main element but that's not the case.