Fix selectors, fill in generateSavedProjects logic
generateSavedProjects function is invoked at page load, and it dynamically populates the sidebar with elements, which have been styled, although I'm not sure how I want to style them to look like links although they won't be links necessarily.
I also declared generateProject however its just a skeleton function for now, this will be invoked when one of the projects in the sidebar has been clicked, so we'll need to dynamically create elements for each projects and thus of course that projects data and hide them like we did with the modal.
I also started testing minLength on the project name input field, however the modal is still saving so I need to look into validation not allowing the function to continue although I could write javascript to disallow it and add pop up validation text. I've added pseudocode to remind me later.
I've also added data-index dataset attributes by using a for loop instead of a for of loop, which is what I was using up until a second ago.