Change name of generating functions
This new name more clearly fits what the function is doing, generateSavedProjects might have gotten confused with the function that generates content in the main element for a project clicked on.
I've also changed the name of that function to viewProject, as that is basically what the function is doing for the user, creating a view of it.
I've also moved, added, and removed some comments as they were just reminders of things to do.