Add logic to disallow short project names
Instead of using css input validation I've used javascript to disallow invoking the function that creates a new project and alert the user with dynamic text and highlighting on the input field which lasts for 3 seconds, and then disappears as we don't want a nuisance perpetually sitting on their screen.
I also added some place holder work on an event listener which captures child nodes being targeted by placing the listener on the ul in the nav.