Add function to generate a project in the main
This function needed multiple other functions declared as well because of Solid principles, however I'm sure I've failed those principles in other regards.
So I also built a function that builds todo items which uses a function that builds checklist items, these items are more complex than the notes items which are just strings in an array, they are objects.