Add functionality to wipe built projects
This will be used when a user wants to generate a different project in the main element than what they have at any point in time.
I will in the future have a default starter page that tells the user how to use the app, and am thinking of showing it at any point when a user clicks a help button or closes a project, which I haven't built yet either.