Change create project button to toggle modal
Instead of having the button just pop up the modal if clicked again it hides the modal and wipes any input.
Created a button to delete all projects, which prompts a user with a confirmation question just in case.
Added index parameter to specific functions that build elements that will need to rely on having an index dataset attribute value to invoke todoApp functions with.
Added more pseudocode as reminders for features.