Change viewProject to remove lastChild
Since I am now using a template modal for the build todo area, we needed to change remove firstChild to remove lastChild, as the prior was deleting our hidden by default build todo modal.
I've also added logic to wipe the built todos area each time a new one is added, which does take more resources however saves me a headache when setting a todoIndex dataset attribute and its appropriate value.
I've also made a function out of the logic of iterating over a projects todos as we need to use it in multiple places now, when loading a project into the view and when repainting todos when adding a new todo.
And I found a correction to last commit, I needed to state grid and not an empty value to display attribute on newProjectModal in the case of a user reclicking create project to toggle off the modal vs showing the modal with one click.