Move newProjectModal to html template
I have omitted the dynamic code that created the newProjectModal from the js and added it into the html template as an aside still and thus it is hidden just like the newTodoModal, and eventually just like the newChecklistModal, and possibly the newNoteModal now that I think of it.
Also made another change to the toggle logic of the create project button, I swear it works now... I was using the not operator incorrectly and am now using the not equal operator instead, I tested it, it works... It wasn't working correctly on refresh of page.