Add ui and logic for todo sections
I got lost in changes and haven't committed in a while, learning experience...
Fixed expression values in the buildTodo function that pull values from a todo object from the todoApp to build a view of that project correctly.
Applied some styling to the built todos to visually set them apart from each other for now.
Added styling to the buildProjectTodos ul to space it's li's (todo items) from the main elements borders.
Worked on logic of createTodoButton listener, which shows the create new todo modal and includes the logic of passing user input to the function. Still working on this section...
Also added text to rendered elements of the todos to help users understand what data they are looking at, for example todo status before the value of the todos complete property value rendered.
Added delete buttons next to each note of a todo so users can delete notes if they need to.
Added explanation text before rendered project name and project description to help the user understand the data they're reading.
Added delete and complete buttons to the todo items, so users can obviously delete a todo if they want or toggle the complete value of the todo if needed, complete strikes through all text in a todo, and obviously not complete removes the strike through.