Add omitted variable which targets project name
Somehow a variable was deleted that we used to target the project name input label on the new project modal, which resulted in errors when a user would attempt to use a project name less than 3 characters long instead of alerting the user like it worked before.
Also added logic for alerting user that their new todo name length is not long enough, it's pretty similar to the new project alert, I just noticed that these aren't limited to one occurrence, if a user clicks multiple times then this will create multiple warnings, I'm not going to worry about it for now.
I've also added some styling to the input field when it's in this invalid state via the js.