Move variable initializations atop, fix etc
Moved all variable initializations to the top of the file to keep track of them more easily, I couldn't have done this before I moved everything to the html template because dynamic elements were being created within other scopes/etc.
Added logic for newNoteModalCancelButton to wipe the input field in the newNoteModal input field.
Added some other elements for adding text and buttons to note elements, not sure if needed but I think there was a problem with adding innerText and appending a button.
I have a lot of things I'm currently stuck on and I think they all hinge on if methods for todos and checklist objects are actually being stringified, and I don't think they are.