Change loadStorage logic, as it's wrong
The past code was wrong in that I was trying to swap out objects for class instances but how I went about it does not work because you can't deep copy an object by simply assigning pointers to a new variable, so I've gone ahead and made changes to use the object to create a class instance and then use splice method to delete the starting object and replace it with the class instance object, this has thwarted me for multiple days but I'm so happy to have figured it out with no help even though help would have shortened the amount of time spent.
Also renamed some variables in the function to better reflect what they are holding/pointing to.