Fix expression used in addNote event listener
Removed attributes property when attempting to set dataset property on a checklist dom element, attributes is not needed to set dataset attribute.
Returned checklistItem in the buildChecklist function, this is called in a for loop that appends the return of this function into a ul, however the function wasn't returning anything.
Added an invocation of viewProject into the listener of the deleteNoteButton as we will want to rerender the project when notes are removed from it.
I'm still having problems with creating/displaying notes, the delete button is rendered but not the note text which should be inserted into a span which is inserted into a list item which symbolizes the note as a whole object.
Also added an invocation of viewProject when a new note is added as we want to render that note after they add it, and added a wipe of the input field at that time as well.
If I use a premade project, the notes are rendered correctly, but if i try to add them myself they are not beign displayed and I think that means they are not being created into the js correctly from the UI.
Omitted value expression from variable pointer to the input field of a new note, as it was saving value on javascript load, which is not when we want it to pull the value but when event listeners fire.