Fix logic when deleting projects, add explainer
Prior to these changes when a user deleted all projects the create new project modal would not appear when asked for, this was because it was being removed from the dom, which I've fixed by checking the amount of projects at the time of deleting all and then wiping if there are any otherwise do nothing except draw the explainer.
Added the drawExplainer function which is a welcome note to users on the page about what this project is and how they can use it. I might expand this later to hold more info.