Add logic for side nav project delete buttons
The logic here when any of these buttons are clicked allows a user to delete a project from the array of projects of the app, and then it causes the side nav to be repainted which resets up the listeners on those newly created buttons.
This could have been written better by delegating a listener down from the ul through the li's to the buttons, but I'm really just trying to get this project done and have a lot more necessary things I'd like to just finish.