Rename button, evaluate current content generation
Added a toggle ability to the add note button, I guess I forgot to do that before but we definitely want users to be able to close the modal by clicking the same button.
Changed modalCancelButton to newProjectModalCancelButton for better clarity.
I think the pre-made modals are not meeting the needs I hoped they would, I didn't even think that about the event listeners not overwriting past listeners but are stacking, i hoped that when i generated a view for a project the listeners on the modals would be swapped out but that is not the case, so I think I need to dynamically generate modals for specifics, for example generate a newTodoModal for a specific project, generate a newNoteModal for a specific todo, generate a newChecklistModal for a specific todo, and hopefully this way they are tied to the projects, to the todos, but I'm still thinking that index values are going to be needed everywhere...