Add changeProjectTitle button and logic
This entailed creating a modal that allows us to let users change project titles after they've been created, which consists of an input field and some buttons like saving their changes or canceling this decision.
I've also copied over some validation for the input for the length of the new title, and have changed it to work here, when the length is greater than 2 then it changes the stored project title, but if what they enter is less than 3 then they are told by the app below the input field that their input is not long enough for 3 seconds and then that message is removed.