Move the toggleTodo from Project to Todo class
I also made the change to the UIController  to call the method from the todo and not the project, and this works. Prior to this change I was trying to call this method on the project itself but it kept telling me complete wasn't a property... Well it works now and that's all that matters.