Add logic for validation on new project name
Added validation that will not allow a new project to be made if that project name is less than 3 characters long, and displays a message to the user which disappears fater 3 seconds.
I'm either done with this project or will add ways to sort projects based on duedates and priorty values of todo items, and ways to sort todos within a viewed project, however I've been working on this rpoject for too long would love to move on.