Add styling
To style the project araea to sticky when todos are scrolled I needed to create a new section element to hold the project title and description containers.
I just realized that my add todo button gets scrolled up and maybe I should position it in a way that it's always in view.