Add styles and edit js to alter styiling
I made various changes here to style the ui , including moving the 'add todo' button when their are no todos on a project.
I have yet to make the sorting functionality that I previously mentioned and won't be creating it today for sure, and don't think I'll make it for this project unless I come back later to it.