Correct misspelt words, remove unnecessary code
I've gone through and corrected misspelt words and removed code that didn't need to exist, like wiping input fields when user successfully creates a new project, as we just remove the modal anyways.
I've also noticed that when a user toggles a todo, that change is not saved when reloading the page, which is pulling from the localStorage I believe, I've tested many different things and even changed the value from boolean to string and still cannot get the changed value to persist through localStorage, I've even created a Project version of todoToggle and tested with it only to come up with the same result... I can get it to change locally which means it's saving correctly in the todoApp.projects array, but it will not change in localStorage.