Fix complete argument to Todo constructor
The problem is still in between saving to projects and then into localStorage, when I reload the local projects todo is holding a 'false' value to the complete property while the localStorage version is holding a 'true'....
The problem was in loadStorage, when we remake instances of our todos I didn't have the stored value for complete pass into the constructor for the Todo class, fixing this fixed the problem.
Also changed the spelling of some checklist items, only because formatter was poking at me about their names.